--- Mincho_The_Mint_Choco_Slime.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28729.10
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mincho_Infection", "Mincho_The_Mint_Choco_Slime\Mincho_Infection.csproj", "{666D2CAB-A18D-4C6E-957D-1C7472643A65}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{666D2CAB-A18D-4C6E-957D-1C7472643A65}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{666D2CAB-A18D-4C6E-957D-1C7472643A65}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{666D2CAB-A18D-4C6E-957D-1C7472643A65}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{666D2CAB-A18D-4C6E-957D-1C7472643A65}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2ADC1FFA-AB84-4A8A-939A-18C334D390AB}
-	EndGlobalSection
-EndGlobal
