--- Mincho_The_Mint_Choco_Slime/Properties/AssemblyInfo.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Diagnostics;
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyTitle("Mincho_The_Mint_Choco_Slime")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Mincho_The_Mint_Choco_Slime")]
-[assembly: AssemblyCopyright("Copyright ©  2019")]
-[assembly: AssemblyTrademark("")]
-[assembly: ComVisible(false)]
-[assembly: Guid("2b33ad12-50f7-4084-ad98-2cbc42bca003")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
