+++ Source/RW_Mincho.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30002.166
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RW_Mincho", "RW_Mincho.csproj", "{06004C7E-19CA-467E-BC45-B91FAEBE1CB8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{06004C7E-19CA-467E-BC45-B91FAEBE1CB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06004C7E-19CA-467E-BC45-B91FAEBE1CB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06004C7E-19CA-467E-BC45-B91FAEBE1CB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06004C7E-19CA-467E-BC45-B91FAEBE1CB8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0C525439-5C11-4D99-8439-94309825442A}
+	EndGlobalSection
+EndGlobal
